RayOfLight
A ray of light comes in from the upper left corner of an MxN sized
window, with 45° angled edges. The ray is reflected when it reaches the
first or the last line, or the first or last column respectively.
Display all the positions reached until the ray travels to a window
corner.